requirements.txt
# Production dependencies for Synapse

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Storage & Caching
redis>=5.0.1

# Machine Learning & Tracking
mlflow>=2.9.0

# External API Clients
slack-sdk>=3.26.0
PyGithub>=2.1.1
openai>=1.0.0

# Configuration & Validation
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.24.0

# Async Support
asyncio-mqtt>=0.16.0

# Logging
structlog>=23.0.0

# Monitoring
prometheus-client>=0.16.0

# Security
cryptography>=41.0.0

# Performance
orjson>=3.9.0